+++ search-films/src/components/shared/CssUtils.js
@@ -2,5 +2,9 @@ class CssUtils {
 	static addPxToCssProperty(property) {
 		return property + 'px';
 	}
+	static addClassToDefaultClassName(defaultClassName, addedClass) {
+		return defaultClassName + (addedClass ? ' ' + addedClass : '');
+	}
+
 }
 export default CssUtils;